+++ R/fl_DownloadFjord.R
@@ -6,11 +6,15 @@
 #'
 #' @param fjord Expects a character vector for one of the 8 available fjords.
 #' See \code{\link{fl_ListFjords}} for the list of possible choices.
-#' @param monthly The layer of monthly data the user wants to download. The default "PAR_B"
-#' will download monthly bottom PAR data. The other option "K_PAR" will download monthly
-#' values for the light extinction coefficient (i.e.  K_PAR) in the water column.
+#' @param layer The layer of data the user wants to download. The default "PAR_B"
+#' will download monthly bottom PAR data, "K_PAR" will download monthly
+#' values for the light extinction coefficient (i.e. K_PAR) in the water column,
+#' "ClimSD" will download the standard deviations for the monthly climatologies, and
+#' "YearlySD" will download the standard deviations for the yearly climatologies.
 #' Note that if monthly K_PAR data are chosen, the file will be saved as e.g.
-#' "kong_MonthlyKpar.nc", whereas PAR_B data will be saved simply as e.g. "kong.nc"
+#' "kong_MonthlyKpar.nc". "ClimSD" and "YearlySD" data will also have this character
+#' string appended to the file name. Whereas PAR_B data will be saved simply as e.g.
+#' "kong.nc". NB: Only the "PAR_B" data contain the full list of metadata variables.
 #' @param dirdata The directory where the user would like to download the data.
 #'
 #' @return The downloaded NetCDF file contains the following variables:
@@ -35,13 +39,15 @@
 #' }
 #'
 fl_DownloadFjord <- function(fjord,
-                             monthly = "PAR_B",
+                             layer = "PAR_B",
                              dirdata = NULL) {
   opt_orig <- options()
   on.exit(options(opt_orig))
   options(timeout = 0)
-	urlobsvlfr <- "ftp://ftp.obs-vlfr.fr/pub/gentili/NC_c2_Fjords_MonthlyKpar"
 	urlpangaea <- "https://download.pangaea.de/dataset/962895/files"
+	urlobsvlfr_Kpar <- "ftp://ftp.obs-vlfr.fr/pub/gentili/NC_c2_Fjords_MonthlyKpar"
+	urlobsvlfr_ClimSD <- "ftp://ftp.obs-vlfr.fr/pub/gentili/NC_c2_Fjords_ClimSD"
+	urlobsvlfr_YearlySD <- "ftp://ftp.obs-vlfr.fr/pub/gentili/NC_c2_Fjords_YearlySD"
 	dlnote <- "Please check your internet connection."
 	if(curl::has_internet()){
 	  fjords <- fl_ListFjords()
@@ -50,14 +56,20 @@ fl_DownloadFjord <- function(fjord,
 	  }
 	  if(is.null(dirdata)) stop("Please provide the pathway to where you would like to download the data.")
 	  if(! file.exists(dirdata)) stop("Please ensure that the chosen directory exists.")
-	  if(monthly == "PAR_B"){
+	  if(layer == "PAR_B"){
 	    ncfile <- paste(fjord, "nc", sep = ".")
 	    ncurl <- urlpangaea
-	  } else if(monthly == "K_PAR"){
+	  } else if(layer == "K_PAR"){
 	    ncfile <- paste0(fjord,"_MonthlyKpar.nc")
-	    ncurl <- urlobsvlfr
+	    ncurl <- urlobsvlfr_Kpar
+	  } else if(layer == "ClimSD"){
+	    ncfile <- paste0(fjord,"_ClimSD.nc")
+	    ncurl <- urlobsvlfr_ClimSD
+	  } else if(layer == "YearlySD"){
+	    ncfile <- paste0(fjord,"_YearlySD.nc")
+	    ncurl <- urlobsvlfr_YearlySD
 	  } else {
-	    stop("Please ensure the 'monthly' value is either 'PAR_B' or 'K_PAR'")
+	    stop("Please ensure the 'layer' value is either 'PAR_B', 'K_PAR', 'ClimSD', or 'YearlySD'")
 	  }
 	  localf <- paste(dirdata, ncfile, sep = "/")
 	  if(! file.exists(localf)) {