+++ R/fl_LoadFjord.R
@@ -5,11 +5,13 @@
 #'
 #' @param fjord Expects a character vector for one of the 8 available fjords.
 #' See \code{\link{fl_ListFjords}} for the list of possible choices.
-#' @param monthly The layer of monthly data the user wants to load. The default "PAR_B"
-#' will load monthly bottom PAR data. The other option "K_PAR" will load monthly
-#' values for the light extinction coefficient (i.e.  K_PAR) in the water column.
-#' Note that the K_PAR files are a slimmed down version of the PAR_B files and do not
-#' contain global values.
+#' @param layer The layer of data the user wants to load. The default "PAR_B"
+#' will load monthly bottom PAR data. The other options, "K_PAR" will load monthly
+#' values for the light extinction coefficient (i.e.  K_PAR) in the water column,
+#' "ClimSD" will load the standard deviation values for the monthly climatologies,
+#' and "YearlySD" will load the standard deviation values for the yearly climatologies.
+#' Note that only the PAR_B files and do not contain all global values and metadata.
+#' The other files are in supplement to the PAR_B files.
 #' @param dirdata The directory where the user would like to load the data from.
 #' @param TS The default, \code{FALSE}, will prevent this function from loading the
 #' monthly bottom PAR values. Instead it will load all global, annual, and monthly
@@ -36,21 +38,25 @@
 #' fjorddata_full <- fl_LoadFjord(fjord_code,
 #'                                dirdata = system.file("extdata", package = "FjordLight"), TS = TRUE)
 #'
-fl_LoadFjord <- function(fjord, monthly = "PAR_B", dirdata = NULL, TS = FALSE, verbose = FALSE) {
+fl_LoadFjord <- function(fjord, layer = "PAR_B", dirdata = NULL, TS = FALSE, verbose = FALSE) {
   if(is.null(dirdata)) stop("Please provide the pathway from where you would like to load the data.")
   if(! file.exists(dirdata)) stop("Please ensure that the chosen directory exists.")
-  if(monthly == "PAR_B"){
+  if(layer == "PAR_B"){
     ncfile <- paste0(dirdata,"/",fjord,".nc")
-  } else if(monthly == "K_PAR"){
+  } else if(layer == "K_PAR"){
     ncfile <- paste0(dirdata,"/",fjord,"_MonthlyKpar.nc")
-  } else {
-    stop("Please ensure the 'monthly' value is either 'PAR_B' or 'K_PAR'")
+  } else if(layer == "ClimSD"){
+    ncfile <- paste0(dirdata,"/",fjord,"_ClimSD.nc")
+  } else if(layer == "YearlySD"){
+    ncfile <- paste0(dirdata,"/",fjord,"_YearlySD.nc")
+  } else{
+    stop("Please ensure the 'layer' value is either 'PAR_B', 'K_PAR', 'ClimSD', or 'YearlySD'")
   }
 	nc <- ncdf4::nc_open(ncfile, verbose = verbose)
 	dims <- names(nc$dim)
 	vars <- names(nc$var)
-	if(!TS & monthly == "PAR_B") vars <- vars[! vars %in% "MonthlyPARbottom"]
-	if(!TS & monthly == "K_PAR") vars <- vars[! vars %in% "MonthlyKpar"]
+	if(!TS & layer == "PAR_B") vars <- vars[! vars %in% "MonthlyPARbottom"]
+	if(!TS & layer == "K_PAR") vars <- vars[! vars %in% "MonthlyKpar"]
 	vars_attributes <- list()
 	for(d in dims) {
 		assign(d, ncdf4::ncvar_get(nc, d))