+++ tests/testthat/test-fl_LoadFjord.R
@@ -5,8 +5,8 @@ test_that("loading errors are correct", {
                "Please provide the pathway from where you would like to load the data.")
   expect_error(fl_LoadFjord("test", dirdata = "guava"),
                "Please ensure that the chosen directory exists.")
-  expect_error(fl_LoadFjord("test", dirdata = system.file("extdata", package = "FjordLight"), monthly = "banana"),
-               "Please ensure the 'monthly' value is either 'PAR_B' or 'K_PAR'")
+  expect_error(fl_LoadFjord("test", dirdata = system.file("extdata", package = "FjordLight"), layer = "banana"),
+               "Please ensure the 'layer' value is either 'PAR_B', 'K_PAR', 'ClimSD', or 'YearlySD'")
 })
 
 test_that("loading with and without TS works", {
@@ -26,4 +26,18 @@ test_that("loading monthly K_PAR works", {
   test_TS <- fl_LoadFjord("test", "K_PAR", dirdata = system.file("extdata", package = "FjordLight"), TS = TRUE)
   expect_type(test_TS, "list")
   expect_type(test_TS$MonthlyKpar, "double")
+})
+
+test_that("loading ClimSD works", {
+  test_TS <- fl_LoadFjord("test", "ClimSD", dirdata = system.file("extdata", package = "FjordLight"), TS = TRUE)
+  expect_type(test_TS, "list")
+  expect_equal(length(test_TS$glob_attributes), 0)
+  expect_type(test_TS$ClimPARbottomSD, "double")
+})
+
+test_that("loading YearlySD works", {
+  test_TS <- fl_LoadFjord("test", "YearlySD", dirdata = system.file("extdata", package = "FjordLight"), TS = TRUE)
+  expect_type(test_TS, "list")
+  expect_equal(length(test_TS$glob_attributes), 0)
+  expect_type(test_TS$YearlyPARbottomSD, "double")
 })